+++ ChANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Commit 06-09-2021 (v2.4.2)
+
+- Add `Whoops\Handler\PrettyPageHandler` to deal with AJAX requests with an equally-informative JSON response.
+
 ## Commit 06-09-2021 (v2.4.0)
 
 - Suppression de la librairie `Horizom\Auth` qui permettait de gérer l'authentification